up examples

+++ swich/code/examples/types.c
@@ -1,5 +1,5 @@
 void main() {
-    short a = 1;
+    int64 a = 1;
     int b = 2;
     long c = 3, d = 4;
     print(a + b + c + d);
